Module google_dns1::api

Structs

Change
A Change represents a set of ResourceRecordSet additions and deletions applied atomically to a ManagedZone. ResourceRecordSets within a ManagedZone are modified by creating a new Change element in the Changes collection. In turn the Changes collection also records the past modifications to the ResourceRecordSets in a ManagedZone. The current state of the ManagedZone is the sum effect of applying all Change elements in the Changes collection in sequence.

ChangeCreateCall
Atomically updates the ResourceRecordSet collection.

ChangeGetCall
Fetches the representation of an existing Change.

ChangeListCall
Enumerates Changes to a ResourceRecordSet collection.

ChangeMethods
A builder providing access to all methods supported on change resources. It is not used directly, but through the Dns hub.

ChangesListResponse
The response to a request to enumerate Changes to a ResourceRecordSets collection.

Dns
Central instance to access all Dns related resource activities

DnsKey
A DNSSEC key pair.

DnsKeyDigest
There is no detailed description.

DnsKeyGetCall
Fetches the representation of an existing DnsKey.

DnsKeyListCall
Enumerates DnsKeys to a ResourceRecordSet collection.

DnsKeyMethods
A builder providing access to all methods supported on dnsKey resources. It is not used directly, but through the Dns hub.

DnsKeySpec
Parameters for DnsKey key generation. Used for generating initial keys for a new ManagedZone and as default when adding a new DnsKey.

DnsKeysListResponse
The response to a request to enumerate DnsKeys in a ManagedZone.

ManagedZone
A zone is a subtree of the DNS namespace under one administrative responsibility. A ManagedZone is a resource that represents a DNS zone hosted by the Cloud DNS service.

ManagedZoneCreateCall
Creates a new ManagedZone.

ManagedZoneDeleteCall
Deletes a previously created ManagedZone.

ManagedZoneDnsSecConfig
There is no detailed description.

ManagedZoneForwardingConfig
There is no detailed description.

ManagedZoneForwardingConfigNameServerTarget
There is no detailed description.

ManagedZoneGetCall
Fetches the representation of an existing ManagedZone.

ManagedZoneListCall
Enumerates ManagedZones that have been created but not yet deleted.

ManagedZoneMethods
A builder providing access to all methods supported on managedZone resources. It is not used directly, but through the Dns hub.

ManagedZoneOperationGetCall
Fetches the representation of an existing Operation.

ManagedZoneOperationListCall
Enumerates Operations for the given ManagedZone.

ManagedZoneOperationMethods
A builder providing access to all methods supported on managedZoneOperation resources. It is not used directly, but through the Dns hub.

ManagedZoneOperationsListResponse
There is no detailed description.

ManagedZonePatchCall
Applies a partial update to an existing ManagedZone.

ManagedZonePeeringConfig
There is no detailed description.

ManagedZonePeeringConfigTargetNetwork
There is no detailed description.

ManagedZonePrivateVisibilityConfig
There is no detailed description.

ManagedZonePrivateVisibilityConfigNetwork
There is no detailed description.

ManagedZoneReverseLookupConfig
There is no detailed description.

ManagedZoneServiceDirectoryConfig
Contains information about Service Directory-backed zones.

ManagedZoneServiceDirectoryConfigNamespace
There is no detailed description.

ManagedZoneUpdateCall
Updates an existing ManagedZone.

ManagedZonesListResponse
There is no detailed description.

Operation
An operation represents a successful mutation performed on a Cloud DNS resource. Operations provide: - An audit log of server resource mutations. - A way to recover/retry API calls in the case where the response is never received by the caller. Use the caller specified client_operation_id.

OperationDnsKeyContext
There is no detailed description.

OperationManagedZoneContext
There is no detailed description.

PoliciesListResponse
There is no detailed description.

PoliciesPatchResponse
There is no detailed description.

PoliciesUpdateResponse
There is no detailed description.

Policy
A policy is a collection of DNS rules applied to one or more Virtual Private Cloud resources.

PolicyAlternativeNameServerConfig
There is no detailed description.

PolicyAlternativeNameServerConfigTargetNameServer
There is no detailed description.

PolicyCreateCall
Creates a new Policy.

PolicyDeleteCall
Deletes a previously created Policy. Fails if the policy is still being referenced by a network.

PolicyGetCall
Fetches the representation of an existing Policy.

PolicyListCall
Enumerates all Policies associated with a project.

PolicyMethods
A builder providing access to all methods supported on policy resources. It is not used directly, but through the Dns hub.

PolicyNetwork
There is no detailed description.

PolicyPatchCall
Applies a partial update to an existing Policy.

PolicyUpdateCall
Updates an existing Policy.

Project
A project resource. The project is a top level container for resources including Cloud DNS ManagedZones. Projects can be created only in the APIs console. Next tag: 7.

ProjectGetCall
Fetches the representation of an existing Project.

ProjectManagedZoneRrsetCreateCall
Creates a new ResourceRecordSet.

ProjectManagedZoneRrsetDeleteCall
Deletes a previously created ResourceRecordSet.

ProjectManagedZoneRrsetGetCall
Fetches the representation of an existing ResourceRecordSet.

ProjectManagedZoneRrsetPatchCall
Applies a partial update to an existing ResourceRecordSet.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Dns hub.

Quota
Limits associated with a Project.

ResourceRecordSet
A unit of data that is returned by the DNS servers.

ResourceRecordSetListCall
Enumerates ResourceRecordSets that you have created but not yet deleted.

ResourceRecordSetMethods
A builder providing access to all methods supported on resourceRecordSet resources. It is not used directly, but through the Dns hub.

ResourceRecordSetsDeleteResponse
There is no detailed description.

ResourceRecordSetsListResponse
There is no detailed description.

ResponseHeader
Elements common to every response.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.