Struct google_dns1::Quota

pub struct Quota {
    pub rrsets_per_managed_zone: Option<i32>,
    pub kind: Option<String>,
    pub rrset_additions_per_change: Option<i32>,
    pub total_rrdata_size_per_change: Option<i32>,
    pub resource_records_per_rrset: Option<i32>,
    pub rrset_deletions_per_change: Option<i32>,
    pub managed_zones: Option<i32>,
}

Limits associated with a Project.

This type is not used in any activity, and only used as part of another schema.

Fields

rrsets_per_managed_zone

Maximum allowed number of ResourceRecordSets per zone in the project.

kind

Identifies what kind of resource this is. Value: the fixed string "dns#quota".

rrset_additions_per_change

Maximum allowed number of ResourceRecordSets to add per ChangesCreateRequest.

total_rrdata_size_per_change

Maximum allowed size for total rrdata in one ChangesCreateRequest in bytes.

resource_records_per_rrset

Maximum allowed number of ResourceRecords per ResourceRecordSet.

rrset_deletions_per_change

Maximum allowed number of ResourceRecordSets to delete per ChangesCreateRequest.

managed_zones

Maximum allowed number of managed zones in the project.