Struct google_dns1::ManagedZone

pub struct ManagedZone {
    pub kind: Option<String>,
    pub description: Option<String>,
    pub name_servers: Option<Vec<String>>,
    pub creation_time: Option<String>,
    pub id: Option<String>,
    pub dns_name: Option<String>,
    pub name_server_set: Option<String>,
    pub name: Option<String>,
}

A zone is a subtree of the DNS namespace under one administrative responsibility. A ManagedZone is a resource that represents a DNS zone hosted by the Cloud DNS service.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get managed zones (response)

list managed zones (none)

create managed zones (request|response)

delete managed zones (none)

Fields

kind

Identifies what kind of resource this is. Value: the fixed string "dns#managedZone".

description

A mutable string of at most 1024 characters associated with this resource for the user's convenience. Has no effect on the managed zone's function.

name_servers

Delegate your managed_zone to these virtual name servers; defined by the server (output only)

creation_time

The time that this resource was created on the server. This is in RFC3339 text format. Output only.

id

Unique identifier for the resource; defined by the server (output only)

dns_name

The DNS name of this managed zone, for instance "example.com.".

name_server_set

Optionally specifies the NameServerSet for this ManagedZone. A NameServerSet is a set of DNS name servers that all host the same ManagedZones. Most users will leave this field unset.

name

User assigned name for this resource. Must be unique within the project. The name must be 1-32 characters long, must begin with a letter, end with a letter or digit, and only contain lowercase letters, digits or dashes.